# api/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Tapis Files API
  description: My API
  contact:
    name: CicSupport
    url: http://tacc.utexas.edu
    email: cicsupport@tacc.utexas.edu
  license:
    name: Apache 2.0
    url: http://foo.bar
  version: "0.0"
servers:
- url: http://localhost:8080/
  description: localhost
  variables: {}
- url: https://dev.develop.tapis.io
  description: development
  variables: {}
security:
- Bearer: []
tags:
- name: file operations
- name: share
- name: permissions
- name: transfers
paths:
  /v3/files/content/{systemId}/{path}:
    get:
      tags:
      - content
      summary: Retrieve a file from the files service
      description: Get file contents/serve file
      operationId: filesGetContents
      parameters:
      - name: systemId
        in: path
        description: System ID
        required: true
        schema:
          type: string
        example: system123
      - name: path
        in: path
        description: File path
        required: true
        schema:
          type: string
        example: /folderA/folderB/
      - name: range
        in: header
        description: Range of bytes to send
        schema:
          $ref: '#/components/schemas/HeaderByteRange'
        example: range=0,999
      - name: more
        in: header
        description: Send 1k of UTF-8 encoded string back starting at 'page' 1, ex
          more=1
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: OK
        401:
          description: Not Authenticated
        404:
          description: Not Found
        403:
          description: Not Authorized
  /v3/files/healthcheck:
    get:
      tags:
      - health
      summary: Health check
      description: Health check
      operationId: healthCheck
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheckResponse'
  /v3/files/ops/{systemId}/{path}:
    get:
      tags:
      - file operations
      summary: List files/objects in a storage system.
      description: List files in a bucket
      operationId: listFiles
      parameters:
      - name: systemId
        in: path
        description: System ID
        required: true
        schema:
          type: string
        example: system123
      - name: path
        in: path
        description: path relative to root of bucket/folder
        required: true
        schema:
          type: string
        example: /folderA/folderB/
      - name: limit
        in: query
        description: pagination limit
        schema:
          type: integer
          format: int32
          default: 1000
        example: 100
      - name: offset
        in: query
        description: pagination offset
        schema:
          type: integer
          format: int64
          default: 0
        example: 1000
      - name: meta
        in: query
        description: Return metadata also? This will slow down the request.
        schema:
          type: boolean
      responses:
        200:
          description: A list of files
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileListingResponse'
        401:
          description: Not Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileStringResponse'
        403:
          description: Not Authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileStringResponse'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileStringResponse'
        500:
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileStringResponse'
    put:
      tags:
      - file operations
      summary: Rename a file or folder
      description: Move/Rename a file in {systemID} at path {path}.
      operationId: rename
      parameters:
      - name: systemId
        in: path
        description: System ID
        required: true
        schema:
          type: string
      - name: path
        in: path
        description: File path
        required: true
        schema:
          type: string
      - name: newName
        in: query
        description: The new name of the file/folder
        required: true
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileStringResponse'
        401:
          description: Not Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileStringResponse'
        403:
          description: Not Authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileStringResponse'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileStringResponse'
        500:
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileStringResponse'
    post:
      tags:
      - file operations
      summary: Create a directory
      description: Create a directory in the system at path the given path
      operationId: mkdir
      parameters:
      - name: systemId
        in: path
        description: System ID
        required: true
        schema:
          type: string
      - name: path
        in: path
        description: Path
        required: true
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FileStringResponse'
        401:
          description: Not Authenticated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FileStringResponse'
        403:
          description: Not Authorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FileStringResponse'
        500:
          description: Internal Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FileStringResponse'
    delete:
      tags:
      - file operations
      summary: Delete a file or folder
      description: Delete a file in {systemID} at path {path}.
      operationId: delete
      parameters:
      - name: systemId
        in: path
        description: System ID
        required: true
        schema:
          type: string
      - name: path
        in: path
        description: File path
        required: true
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileStringResponse'
        401:
          description: Not Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileStringResponse'
        403:
          description: Not Authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileStringResponse'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileStringResponse'
        500:
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileStringResponse'
  /v3/files/permissions/{systemId}/{path}:
    get:
      tags:
      - permissions
      summary: Get the API user's permissions on a file or folder.
      description: Get the permissions for the API user for the system and path.
      operationId: permissionsSystemIdPathGet
      parameters:
      - name: systemId
        in: path
        description: System ID
        required: true
        schema:
          type: string
      - name: path
        in: path
        description: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: FilePermission
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FilePermissionResponse'
    post:
      tags:
      - permissions
      summary: 'Add permissions on an object. '
      description: 'Add a user to a file/folder. QUESTION - who should be able to
        add permissions? Only the owner? '
      operationId: permissionsSystemIdPathPost
      parameters:
      - name: systemId
        in: path
        description: System ID
        required: true
        schema:
          type: string
      - name: path
        in: path
        description: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: FilePermission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilePermissionResponse'
    delete:
      tags:
      - permissions
      summary: 'Remove permissions on an object for a user. '
      description: 'Remove user permissions to a file/folder. QUESTION - who should
        be able to delete permissions? Only the owner? '
      operationId: permissionsSystemIdPathDelete
      parameters:
      - name: systemId
        in: path
        description: System ID
        required: true
        schema:
          type: string
      - name: path
        in: path
        description: path
        required: true
        schema:
          type: string
      - name: username
        in: query
        description: Username to remove
        required: true
        schema:
          type: string
      responses:
        200:
          description: FilePermission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilePermissionResponse'
  /v3/files/share/{systemId}/{path}:
    get:
      tags:
      - share
      summary: 'List the shares on a file resource. '
      description: 'List all shares on a given file resource. '
      operationId: shareList
      parameters:
      - name: systemId
        in: path
        description: System ID
        required: true
        schema:
          type: string
      - name: path
        in: path
        description: System ID
        required: true
        schema:
          type: string
      responses:
        200:
          description: List of shares
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SharedFileObject'
    post:
      tags:
      - share
      summary: 'Grant temporary access to a file resource. '
      description: 'Creates a link that is valid for the requested validity time for
        the given user for the resource in {systemId} at path {path} '
      operationId: shareFile
      parameters:
      - name: systemId
        in: path
        description: System ID
        required: true
        schema:
          type: string
      - name: path
        in: path
        description: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShareFileRequest'
      responses:
        200:
          description: Shared file object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedFileObject'
    delete:
      tags:
      - share
      summary: 'Revoke a shared file resource '
      description: 'Removes any outstanding shares on a file resource. '
      operationId: shareDelete
      parameters:
      - name: systemId
        in: path
        description: System ID
        required: true
        schema:
          type: string
      - name: path
        in: path
        description: System ID
        required: true
        schema:
          type: string
      responses:
        200:
          description: Shared file object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedFileObject'
  /v3/files/transfers:
    post:
      tags:
      - transfers
      summary: Transfer data
      description: This creates a background task which will transfer files into the
        storage system
      operationId: createTransferTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferTaskRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferTaskResponse'
  /v3/files/transfers/{transferTaskId}:
    get:
      tags:
      - transfers
      summary: Get a transfer task
      operationId: getTransferTask
      parameters:
      - name: transferTaskId
        in: path
        description: Transfer task ID
        required: true
        schema:
          type: string
        example: 6491c2a5-acb2-40ef-b2c0-bc1fc4cd7e6c
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferTaskResponse'
    delete:
      tags:
      - transfers
      summary: Stop/Cancel a transfer task
      operationId: cancelTransferTask
      parameters:
      - name: transferTaskId
        in: path
        description: Transfer task ID
        required: true
        schema:
          type: string
        example: 6491c2a5-acb2-40ef-b2c0-bc1fc4cd7e6c
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TapisResponse'
  /v3/files/transfers/{transferTaskId}/history:
    get:
      tags:
      - transfers
      summary: Get history of a transfer task
      operationId: getTransferTaskHistory
      parameters:
      - name: transferTaskId
        in: path
        description: Transfer task ID
        required: true
        schema:
          type: string
          format: uuid
        example: 6491c2a5-acb2-40ef-b2c0-bc1fc4cd7e6c
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferTaskResponse'
components:
  schemas:
    HeaderByteRange:
      type: object
      properties:
        min:
          type: integer
          format: int64
        max:
          type: integer
          format: int64
    HealthCheckResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        result:
          type: string
        version:
          type: string
    FileStringResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        result:
          type: string
        version:
          type: string
    FileInfo:
      type: object
      properties:
        lastModified:
          type: string
          format: date-time
        name:
          type: string
        path:
          type: string
        size:
          type: integer
          description: size in kB
          format: int64
    FileListingResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        result:
          type: array
          items:
            $ref: '#/components/schemas/FileInfo'
        version:
          type: string
    FilePermission:
      type: object
      properties:
        tenantId:
          type: string
        username:
          type: string
        systemId:
          type: string
        path:
          type: string
        permissions:
          type: string
          enum:
          - READ
          - READWRITE
          - ALL
    FilePermissionResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        result:
          $ref: '#/components/schemas/FilePermission'
        version:
          type: string
    SharedFileObject:
      type: object
      properties:
        creator:
          type: string
          description: Username who shared the file/folder
        sharedWith:
          type: string
          description: Username who was granted access
        created:
          type: string
          description: Creation timestamp in UTC
        expiresIn:
          type: integer
          description: Number of seconds in which the share was set to expire.
          format: int32
        url:
          type: string
          description: Link to the shared file.
    ShareFileRequest:
      required:
      - expiresIn
      - username
      type: object
      properties:
        username:
          type: string
          description: The user with which to share
        expiresIn:
          maximum: 604800
          minimum: 1
          type: integer
          description: Time in seconds of expiration. minimum=1, maximum=604800 (1
            week)
          format: int32
    TransferTask:
      type: object
      properties:
        id:
          type: integer
          format: int64
        tenantId:
          type: string
        username:
          type: string
        sourceSystemId:
          type: string
        sourcePath:
          type: string
        destinationSystemId:
          type: string
        destinationPath:
          type: string
        totalBytes:
          type: integer
          format: int64
        bytesTransferred:
          type: integer
          format: int64
        retries:
          type: integer
          format: int32
        finalMessage:
          type: string
        uuid:
          type: string
          description: Unique ID of the task.
          format: uuid
        status:
          type: string
          description: The status of the task, such as ACCEPTED, IN_PROGRESS, COMPLETED,
            CANCELLED
          example: PENDING
        created:
          type: integer
          description: Timestamp in UTC of task creation.
          format: int64
    TransferTaskResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        result:
          $ref: '#/components/schemas/TransferTask'
        version:
          type: string
    TransferTaskRequest:
      required:
      - destinationURI
      - sourceURI
      type: object
      properties:
        sourceURI:
          type: string
          description: Fully qualified URI, such as tapis://{systemID}/{path} or https://myserver.com/path/to/inputs/
        destinationURI:
          type: string
          description: Fully qualified URI to a tapis system, such as tapis://{systemID}/{path}
    TapisResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        result:
          type: object
        version:
          type: string